book_recommendation/requirements.txt
pandas
google-genai
kagglehub
kaggle
